requirements.txt
google-generativeai==0.4.0
PyGithub==2.1.1
requests==2.31.0
